Time: 81 ms (18.45%), Space: 33.3 MB (50.83%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.cpp
@@ -36,6 +36,6 @@ public:
         }
         
         if (ways[n-1].first == LLONG_MAX) return 0;
-        return static_cast<int>(ways[n-1].second);
+        return ways[n-1].second;
     }
 };
